requirements.txt
python-digitalocean==1.9.0
requests==2.10.0